Extract get type and renderer into util
