+++ src/chart/SingleValueChart.tsx
@@ -1,6 +1,8 @@
 
 import React from 'react';
 import { ChartProps } from './Chart';
+
+import { getRecordType } from '../report/RecordProcessing';
 /**
  * Renders Neo4j records as their JSON representation.
  */
@@ -11,9 +13,11 @@ const NeoSingleValueChart = (props: ChartProps) => {
     const color = props.settings && props.settings.color ? props.settings.color : "rgba(0, 0, 0, 0.87)";
     const textAlign = props.settings && props.settings.textAlign ? props.settings.textAlign : "left";
 
-    const value = (records && records[0] && records[0]["_fields"] && records[0]["_fields"][0]) ? records[0]["_fields"][0].toString() : "";
+    const value = (records && records[0] && records[0]["_fields"] && records[0]["_fields"][0]) ? records[0]["_fields"][0] : "";
+    console.log("value : " + value.toString());
     // Parse display from the value - e.g. integer with thousand separators
-    const displayValue = isNaN(parseInt(value, 10)) ? value : value.replace(/\B(?=(\d{3})+(?!\d))/g, " ");
+    const displayValue = getRecordType(value) == "number" ? value.toString().replace(/\B(?=(\d{3})+(?!\d))/g, " ") : value.toString();
+    console.log("displayValue" + displayValue);
     return <div style={{marginTop: marginTop, textAlign: textAlign, marginLeft: "15px", marginRight: "15px"}}>
         <span style={{fontSize: fontSize, color: color}}>{displayValue}</span>
     </div >;
