+++ src/report/RecordProcessing.tsx
@@ -1,5 +1,9 @@
 import _ from 'lodash';
-import { DateTime } from 'neo4j-driver';
+
+import React from 'react';
+import { Chip } from '@material-ui/core';
+import { withStyles } from '@material-ui/core/styles';
+import Tooltip from '@material-ui/core/Tooltip';
 
 const OPTIONAL_FIELD_UNAVAILABLE_IDENTIFIER = "(none)";
 
@@ -284,4 +288,125 @@ export function getRecordType(value) {
 
     // Use string as default type
     return 'string';
+}
+
+/* HELPER FUNCTIONS FOR RENDERING A FIELD BASED ON TYPE */
+const HtmlTooltip = withStyles((theme) => ({
+    tooltip: {
+        color: 'white',
+        fontSize: theme.typography.pxToRem(12),
+        border: '1px solid #fcfffa',
+    },
+}))(Tooltip);
+
+function addDirection(relationship, start) {
+    relationship.direction = (relationship.start.low == start.identity.low);
+    return relationship;
+}
+
+const rightRelationship = "polygon(10px 0%, calc(100% - 10px) 0%, 100% 50%, 100% calc(100% - 50%), calc(100% - 10px) 100%, 0px 100%, 0% calc(100% - 0px), 0% 0px)"
+const leftRelationship = "polygon(10px 0%, calc(100% - 0%) 0%, 100% 10px, 100% calc(100% - 10px), calc(100% - 0%) 100%, 10px 100%, 0% calc(100% - 50%), 0% 50%)"
+
+function RenderNode(value, key = 0) {
+    return <HtmlTooltip key={key + "-" + value.identity} arrow title={<div><b> {value.labels.length > 0 ? value.labels.join(", ") : "Node"}</b><table><tbody>{Object.keys(value.properties).length == 0 ? <tr><td>(No properties)</td></tr> : Object.keys(value.properties).map((k, i) => <tr key={i}><td key={0}>{k.toString()}:</td><td key={1}>{value.properties[k].toString()}</td></tr>)}</tbody></table></div>}>
+        <Chip label={value.labels.length > 0 ? value.labels.join(", ") : "Node"} />
+    </HtmlTooltip>
+}
+
+function RenderRelationship(value, key = 0) {
+    return <HtmlTooltip key={key + "-" + value.identity} arrow title={<div><b> {value.type}</b><table><tbody>{Object.keys(value.properties).length == 0 ? <tr><td>(No properties)</td></tr> : Object.keys(value.properties).map((k, i) => <tr key={i}><td key={0}>{k.toString()}:</td><td key={1}>{value.properties[k].toString()}</td></tr>)}</tbody></table></div>}>
+        <Chip style={{ borderRadius: 0, clipPath: (value.direction == undefined) ? "none" : ((value.direction) ? rightRelationship : leftRelationship) }} label={value.type} />
+    </HtmlTooltip>
+}
+
+function RenderPath(value) {
+    return value.segments.map((segment, i) => {
+        return RenderSubValue((i < value.segments.length - 1) ?
+            [segment.start, addDirection(segment.relationship, segment.start)] :
+            [segment.start, addDirection(segment.relationship, segment.start), segment.end], i)
+    });
+}
+
+function RenderArray(value) {
+    const mapped = value.map((v, i) => {
+        return <div>
+            {RenderSubValue(v)}
+            {i < value.length - 1 && !valueIsNode(v) && !valueIsRelationship(v) ? <span>,&nbsp;</span> : <></>}
+        </div>
+    });
+    return mapped;
+}
+
+function RenderString(value) {
+    const str = value ? value.toString() : "";
+    if (str.startsWith("http") || str.startsWith("https")) {
+        return <a target="_blank" href={str}>{str}</a>;
+    }
+    return str;
+}
+
+function RenderNumber(value) {
+    const number = value.toString().replace(/\B(?=(\d{3})+(?!\d))/g, " ");
+    return number;
+}
+
+function RenderSubValue(value, key = 0) {
+    if (value == undefined) {
+        return "";
+    }
+    const type = getRecordType(value);
+    const columnProperties = rendererForType[type];
+
+    if (columnProperties) {
+        if (columnProperties.renderCell) {
+            return columnProperties.renderCell({ value: value });
+        } else if (columnProperties.valueGetter) {
+            return columnProperties.valueGetter({ value: value });
+        }
+    }
+
+    return RenderString(value);
+}
+
+const rendererForType: any = {
+    "node": {
+        type: 'string',
+        renderValue: (c) => RenderNode(c.value),
+    },
+    "relationship": {
+        type: 'string',
+        renderValue: (c) => RenderRelationship(c.value),
+    },
+    "path": {
+        type: 'string',
+        renderValue: (c) => RenderPath(c.value),
+    },
+    "object": {
+        type: 'string',
+        // valueGetter enables sorting and filtering on string values inside the object
+        valueGetter: (c) => { return JSON.stringify(c.value) },
+    },
+    "array": {
+        type: 'string',
+        renderValue: (c) => RenderArray(c.value),
+    },
+    "string": {
+        type: 'string',
+        renderValue: (c) => RenderString(c.value),
+    },
+    "number": {
+        type: 'number',
+        renderValue: (c) => RenderNumber(c.value)
+    },
+    "null": {
+        type: 'string'
+    },
+    "undefined": {
+        type: 'string'
+    }
+};
+
+export function getRendererForValue(value) {
+    const type = getRecordType(value);
+    return rendererForType[type];
 }
